refactor(storage): move queries module up a level
