+++ storage/src/db/crud/mod.rs
@@ -2,5 +2,4 @@ pub mod album;
 pub mod artist;
 pub mod collection;
 pub mod playlist;
-pub mod queries;
 pub mod song;
