+++ storage/src/db/queries/album.rs
